requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
fal_client==0.7.0
frozenlist==1.6.0
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
httpx-sse==0.4.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-telegram-bot==20.6
requests==2.32.3
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
aiohttp